Update plugins
- WPML 4.6.6 -> 5.0.2, WPML String Translation 3.2.8 -> 5.0.2
- ACFML 2.0.5 -> 5.0.0, CF7 Multilingual 1.2.1 -> 5.0.0,
  WP SEO Multilingual 2.1.0 -> 5.0.1
- Advanced Custom Fields Pro 6.8.2 -> 6.7.1 (downgrade)
- Contact Form 7 6.1.6 -> 6.1.7
- Yoast SEO 27.7 -> 28.5
- Post SMTP 3.9.3 -> 4.0.1
- Classic Editor 1.6.7 -> 1.7.0, Disable Comments 2.7.0 -> 2.9.0,
  Duplicate Post 4.6 -> 4.7, Safe SVG 2.4.0 -> 2.5.1,
  User Switching 1.12.0 -> 1.12.2

+++ wp-content/plugins/sitepress-multilingual-cms/classes/API/Hooks/class-wpml-api-hook-copy-post-to-language.php
@@ -1,11 +1,7 @@
 <?php
 
-/**
- * Class WPML_API_Hook_Copy_Post_To_Language
- */
 class WPML_API_Hook_Copy_Post_To_Language implements IWPML_Action {
 
-	/** @var WPML_Post_Duplication $post_duplication */
 	private $post_duplication;
 
 	public function __construct( WPML_Post_Duplication $post_duplication ) {
@@ -18,6 +14,10 @@ class WPML_API_Hook_Copy_Post_To_Language implements IWPML_Action {
 
 	public function copy_post_to_language( $post_id, $target_language, $mark_as_duplicate ) {
 		$duplicate_post_id = $this->post_duplication->make_duplicate( $post_id, $target_language );
+
+		if ( ! is_int( $duplicate_post_id ) ) {
+			return $duplicate_post_id;
+		}
 
 		if( ! $mark_as_duplicate ) {
 			delete_post_meta( $duplicate_post_id, '_icl_lang_duplicate_of' );
